remove original implementation
The original implementation might have been fast, but
it's very out of date now. Remove it, and also remove
the flags that print type information, as they're not
core functionality and are probably not actually used
by anything.

Also remove the vendor directory and support for
non-module mode, and update various dependencies
to a more recent version.